@@ -1,15 +1,15 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion '25.0.0'
+    compileSdkVersion 29
+    buildToolsVersion '29.0.0'
 
     defaultConfig {
         applicationId "com.gohn.memorize"
-        minSdkVersion 15
-        targetSdkVersion 23
-        versionCode 6
-        versionName "2.0"
+        minSdkVersion 16
+        targetSdkVersion 29
+        versionCode 7
+        versionName "2.1"
         multiDexEnabled true
     }
     buildTypes {
@@ -37,20 +37,30 @@ android {
     }
 }
 
+allprojects {
+    repositories {
+        jcenter()
+    }
+}
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    testCompile 'junit:junit:4.12'
-    compile "com.android.support:multidex:1.0.1"
-    compile 'com.android.support:appcompat-v7:23.1.1'
-    compile 'com.android.support:design:23.1.1'
-    compile 'com.google.code.gson:gson:2.5'
-    compile 'com.github.gabrielemariotti.cards:cardslib-core:2.1.0'
-    compile 'com.github.gabrielemariotti.cards:cardslib-recyclerview:2.1.0'
-    compile 'com.mikepenz:materialdrawer:4.6.3'
-    compile 'com.melnykov:floatingactionbutton:1.3.0'
-    compile 'com.lsjwzh:recyclerviewpager:1.0.11-SNAPSHOT'
-    compile 'gun0912.ted:tedpermission:2.1.0'
-    compile 'net.sourceforge.jexcelapi:jxl:2.6.12'
-    compile files('libs/opencsv-3.3.jar')
-    compile files('libs/aspose-cells-8.4.0.jar')
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation 'androidx.multidex:multidex:2.0.0'
+
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation "androidx.constraintlayout:constraintlayout:2.0.0-beta7"
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
+
+//    implementation 'com.android.support:coordinatorlayout:29.0.0'
+//    implementation 'com.android.support:appcompat-v7:23.1.1'
+//    implementation 'com.android.support:design:29.0.0'
+    implementation 'com.google.code.gson:gson:2.5'
+    implementation 'com.mikepenz:materialdrawer:4.6.3'
+    implementation 'com.melnykov:floatingactionbutton:1.3.0'
+    implementation 'com.lsjwzh:recyclerviewpager:1.0.11-SNAPSHOT'
+    implementation 'gun0912.ted:tedpermission:2.2.3'
+    implementation 'net.sourceforge.jexcelapi:jxl:2.6.12'
+    implementation files('libs/opencsv-3.3.jar')
+    implementation files('libs/aspose-cells-8.4.0.jar')
 }
