@@ -1,7 +1,7 @@
 package com.gohn.memorize.activity.base;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.MenuItem;
 
 public class ActionBarActivity extends AppCompatActivity {
